Allow for more file generators to be configured
This commit introduces a change where the existing file-generator is renamed
"traditional" and as a matter of configuration it's now possible to slot more
kinds of file generators into the system. This is not done in this PR.

REF SMPTNG-6
REF SMPTNG-44
REF SMPTNG-45
